requirements.txt
markdown >= 2.5.2
gitpython >= 0.3
